requirements.txt
kafka-python
yfinance
yahoo_fin
pymongo